M03: move _normalize_genotype from sim.py to hpv.py
PR108 inline #13. Robyn: "very minor comment but I would expect to find
this in hpv.py."

_normalize_genotype lives next to _KNOWN_GENOTYPES in hpv.py now; sim.py
imports it from there. Dropped the now-unused `from .parameters import
genotype_aliases` import in sim.py.

Tests: 121 fast tests pass.
